+++ common/src/main/java/me/rufia/fightorflight/entity/EntityFightOrFlight.java
@@ -4,6 +4,9 @@ import dev.architectury.registry.level.entity.EntityAttributeRegistry;
 import dev.architectury.registry.registries.DeferredRegister;
 import dev.architectury.registry.registries.RegistrySupplier;
 import me.rufia.fightorflight.CobblemonFightOrFlight;
+import me.rufia.fightorflight.entity.projectile.PokemonArrow;
+import me.rufia.fightorflight.entity.projectile.PokemonBullet;
+import me.rufia.fightorflight.entity.projectile.PokemonTracingBullet;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
