+++ common/src/main/java/me/rufia/fightorflight/entity/projectile/AbstractPokemonProjectile.java
@@ -1,6 +1,7 @@
-package me.rufia.fightorflight.entity;
+package me.rufia.fightorflight.entity.projectile;
 
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
+import me.rufia.fightorflight.entity.PokemonAttackEffect;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.syncher.EntityDataAccessor;
 import net.minecraft.network.syncher.EntityDataSerializers;
