+++ common/src/main/java/me/rufia/fightorflight/goals/PokemonRangedAttackGoal.java
@@ -6,6 +6,10 @@ import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
 import me.rufia.fightorflight.CobblemonFightOrFlight;
 import me.rufia.fightorflight.PokemonInterface;
 import me.rufia.fightorflight.entity.*;
+import me.rufia.fightorflight.entity.projectile.AbstractPokemonProjectile;
+import me.rufia.fightorflight.entity.projectile.PokemonArrow;
+import me.rufia.fightorflight.entity.projectile.PokemonBullet;
+import me.rufia.fightorflight.entity.projectile.PokemonTracingBullet;
 import me.rufia.fightorflight.utils.PokemonUtils;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.util.Mth;
