+++ common/src/main/java/me/rufia/fightorflight/mixin/PokemonEntityMixin.java
@@ -2,7 +2,6 @@ package me.rufia.fightorflight.mixin;
 
 
 import com.cobblemon.mod.common.api.moves.Move;
-import com.cobblemon.mod.common.entity.Poseable;
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
 import me.rufia.fightorflight.CobblemonFightOrFlight;
 import me.rufia.fightorflight.PokemonInterface;
@@ -13,7 +12,10 @@ import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.network.syncher.SynchedEntityData;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.entity.*;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.animal.ShoulderRidingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
@@ -21,6 +23,10 @@ import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.util.Arrays;
 
@@ -38,9 +44,10 @@ public abstract class PokemonEntityMixin extends Mob implements PokemonInterface
 
     static {
         DATA_ID_ATTACK_TARGET = SynchedEntityData.defineId(PokemonEntityMixin.class, EntityDataSerializers.INT);
-        ATTACK_TIME=SynchedEntityData.defineId(PokemonEntityMixin.class,EntityDataSerializers.INT);
-        MOVE=SynchedEntityData.defineId(PokemonEntityMixin.class,EntityDataSerializers.STRING);
+        ATTACK_TIME = SynchedEntityData.defineId(PokemonEntityMixin.class, EntityDataSerializers.INT);
+        MOVE = SynchedEntityData.defineId(PokemonEntityMixin.class, EntityDataSerializers.STRING);
     }
+
     protected PokemonEntityMixin(EntityType<? extends ShoulderRidingEntity> entityType, Level level) {
         super(entityType, level);
     }
@@ -58,7 +65,7 @@ public abstract class PokemonEntityMixin extends Mob implements PokemonInterface
                 }
             }
         }
-        return  super.getTarget();
+        return super.getTarget();
     }
 
     public void onSyncedDataUpdated(EntityDataAccessor<?> key) {
@@ -71,49 +78,56 @@ public abstract class PokemonEntityMixin extends Mob implements PokemonInterface
     protected void defineSynchedData() {
         super.defineSynchedData();
         this.entityData.define(DATA_ID_ATTACK_TARGET, 0);
-        this.entityData.define(ATTACK_TIME,0);
-        this.entityData.define(MOVE,"");
+        this.entityData.define(ATTACK_TIME, 0);
+        this.entityData.define(MOVE, "");
     }
 
-    public void setTarget(LivingEntity target){
+    public void setTarget(LivingEntity target) {
         super.setTarget(target);
-        if(target!=null){
-            this.entityData.set(DATA_ID_ATTACK_TARGET,target.getId());
+        if (target != null) {
+            this.entityData.set(DATA_ID_ATTACK_TARGET, target.getId());
         }
     }
+
     @Override
-    public int getAttackTime(){
-         return entityData.get(ATTACK_TIME);
+    public int getAttackTime() {
+        return entityData.get(ATTACK_TIME);
     }
+
     @Override
-    public void setAttackTime(int val){
-        entityData.set(ATTACK_TIME,val);
+    public void setAttackTime(int val) {
+        entityData.set(ATTACK_TIME, val);
     }
+
     @Override
-    public boolean usingBeam(){
-        if(entityData.get(MOVE).equals("")){return false;}
+    public boolean usingBeam() {
+        if (entityData.get(MOVE).equals("")) {
+            return false;
+        }
         return Arrays.stream(CobblemonFightOrFlight.moveConfig().single_beam_moves).toList().contains(getCurrentMove());
     }
+
     @Override
-    public void setCurrentMove(Move move){
-        entityData.set(MOVE,move.getName());
+    public void setCurrentMove(Move move) {
+        entityData.set(MOVE, move.getName());
     }
+
     @Override
-    public String getCurrentMove(){
+    public String getCurrentMove() {
         return entityData.get(MOVE);
     }
-            @Override
-    public InteractionResult mobInteract(Player player, InteractionHand hand) {
 
+    //Don't use Override for this function or you will find that you can't change your pokemon's held item
+    @Inject(method ="mobInteract",at=@At("HEAD"), cancellable = true)
+    private void mobInteractInject(Player player, InteractionHand hand, CallbackInfoReturnable<InteractionResult> cir) {
         ItemStack itemStack = player.getItemInHand(hand);
-        if(itemStack.is(ItemFightOrFlight.POKESTAFF.get())){
-            PokeStaff staff=(PokeStaff) itemStack.getItem();
-            if(staff.canSend(itemStack)){
-                staff.send(player,this,itemStack);
-                return InteractionResult.SUCCESS;
+        if (itemStack.is(ItemFightOrFlight.POKESTAFF.get())) {
+            PokeStaff staff = (PokeStaff) itemStack.getItem();
+            if (staff.canSend(itemStack)) {
+                staff.send(player, this, itemStack);
+                cir.setReturnValue(InteractionResult.SUCCESS);
             }
         }
-        return super.mobInteract(player, hand);
     }
