+++ common/src/main/java/me/rufia/fightorflight/mixin/RendererInjector.java
@@ -3,10 +3,8 @@ package me.rufia.fightorflight.mixin;
 import com.cobblemon.mod.common.client.render.pokemon.PokemonRenderer;
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
 import com.mojang.blaze3d.vertex.PoseStack;
-import me.rufia.fightorflight.CobblemonFightOrFlight;
-import me.rufia.fightorflight.entity.PokemonPostRenderer;
+import me.rufia.fightorflight.client.renderer.PokemonPostRenderer;
 import net.minecraft.client.renderer.MultiBufferSource;
-import net.minecraft.world.entity.Mob;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
